Replacing thread index with port number in checkpoints
Summary: We were using thread index in checkpoints. If the sender & receiver
start ports did not match, or if there was some holes in the list of ports,
this approach would have failed. Replacing thread index with port number. Also
added a test with sender & receiver having different start ports.

Reviewed By: @ldemailly

Differential Revision: D1988322

+++ Protocol.h
@@ -8,7 +8,9 @@
 namespace facebook {
 namespace wdt {
 
-typedef std::pair<int16_t, int64_t> Checkpoint;
+/// checkpoint is a pair of port number and number of successfully transferred
+/// blocks
+typedef std::pair<int32_t, int64_t> Checkpoint;
 
 class Protocol {
  public:
